+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/block/BlockRock.java
@@ -264,7 +264,8 @@ public class BlockRock //'n beats
     SANDSTONE(6, NAME + "_sandstone"),
     WOOD_CHIPS(7, NAME + "_wood_chips"),
     LIMESTONE(8, NAME + "_limestone"),
-    SAND_RED(9, NAME + "_sand_red");
+    SAND_RED(9, NAME + "_sand_red"),
+    SANDSTONE_RED(10, NAME + "_sandstone_red");
 
     private static final EnumType[] META_LOOKUP = Stream.of(EnumType.values())
         .sorted(Comparator.comparing(EnumType::getMeta))